// src/pages/Collaborator/AddEditCollaborator/interfaces.tsx
interface ICollaborator {
  id: string;
  name: string;
  contract: string;
  situation: string;
  start: string;
  admission: string;
  occupation: string;
  computer: string;
  payment: string;
  cost_center: string;
  MEI: string;
  hours: string;
  city: string;
  coordinator: string;
  contact: string;
  birth: string;
  CPF: string;
  identity: string;
  schooling: string;
  mother: string;
  sons: string;
  emergency_contact: string;
  responsible_emergency: string;
  address: string;
  email: string;
  status: string;
  equipment_id: string;
}

export type { ICollaborator };